FIA: renamed FIA views filter plugin, added views_ui filter, updated FIA with filter
